bug fix: redirect before created meeting; feature: /api/init with encryption

+++ src/app/api/init/route.ts
@@ -1,7 +1,11 @@
 import { sql } from "@vercel/postgres";
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 
-export async function GET() {
+export async function POST(req: NextRequest) {
+    let pwd = (await req.json()).pwd;
+    if (pwd !== '12138') {
+        return NextResponse.json({ result: false })
+    }
     await sql`DELETE FROM meetings;`
     await sql`DELETE FROM occupied;`
     await sql`DELETE FROM users_meetings;`;
